Wipe cache OnEnable
Going through a load screen doesn't do a full reloadUI, but adibags handles the changes by calling Disable and then Enable.  OnDisable all the text is hidden, so OnEnable the cache needs to be wiped so that the text can be shown as needed.  This fixes item levels not being show after a loading screen.